four-line-display+rotary: avoid infinite loops
changed two loop counters from byte to uint16_t, to avoid infinite loops - "i < strip.getModeCount()" will always be true when the type of i is too small.

+++ usermods/usermod_v2_rotary_encoder_ui_ALT/usermod_v2_rotary_encoder_ui_ALT.h
@@ -353,7 +353,7 @@ void RotaryEncoderUIUsermod::sortModesAndPalettes() {
 
 byte *RotaryEncoderUIUsermod::re_initIndexArray(int numModes) {
   byte *indexes = (byte *)malloc(sizeof(byte) * numModes);
-  for (byte i = 0; i < numModes; i++) {
+  for (uint16_t i = 0; i < numModes; i++) {              // WLEDMM changed to uint16_t to avoid infinite loop with 16bit mode IDs
     indexes[i] = i;
   }
   return indexes;
@@ -630,7 +630,7 @@ void RotaryEncoderUIUsermod::displayNetworkInfo() {
 void RotaryEncoderUIUsermod::findCurrentEffectAndPalette() {
   if (modes_alpha_indexes == nullptr) return; // WLEDMM bugfix
   currentEffectAndPaletteInitialized = true;
-  for (uint8_t i = 0; i < strip.getModeCount(); i++) {
+  for (uint16_t i = 0; i < strip.getModeCount(); i++) {   // WLEDMM changed to uint16_t to avoid infinite loop with 16bit mode IDs
     if (modes_alpha_indexes[i] == effectCurrent) {
       effectCurrentIndex = i;
       break;
